Migrate fixSpotifyEmbeds; add strong types
