+++ src/plugins/openInApp/index.ts
@@ -20,6 +20,7 @@ import { definePluginSettings } from "@api/Settings";
 import { Devs } from "@utils/constants";
 import definePlugin, { OptionType } from "@utils/types";
 import { showToast, Toasts } from "@webpack/common";
+import { PluginNative } from "main/ipcPlugins";
 import type { MouseEvent } from "react";
 
 const ShortUrlMatcher = /^https:\/\/(spotify\.link|s\.team)\/.+$/;
@@ -44,6 +45,8 @@ const settings = definePluginSettings({
         default: true,
     }
 });
+
+const Native = VencordNative.pluginHelpers.OpenInApp as PluginNative<typeof import("./native")>;
 
 export default definePlugin({
     name: "OpenInApp",
@@ -84,7 +87,7 @@ export default definePlugin({
         if (!IS_WEB && ShortUrlMatcher.test(url)) {
             event?.preventDefault();
             // CORS jumpscare
-            url = await VencordNative.pluginHelpers.OpenInApp.resolveRedirect(url);
+            url = await Native.resolveRedirect(url);
         }
 
         spotify: {
